floats please!

+++ sw/airborne/modules/helicopter/throttle_curve.h
@@ -51,7 +51,7 @@ struct throttle_curve_t {
 
   uint16_t rpm_meas;                          ///< RPM measured
   bool rpm_measured;                          ///< Whenever the RPM is measured
-  int16_t rpm_err_sum;                        ///< Summed RPM error
+  float rpm_err_sum;                          ///< Summed RPM error
   float rpm_fb_p;                             ///< RPM feedback p gain
   float rpm_fb_i;                             ///< RPM feedback i gain
 };
